bot/pyproject.toml
[project]
name = "lms-bot"
version = "0.1.0"
dependencies = [
    "python-telegram-bot>=20.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0",
]